+++ src/main/java/roomescape/domain/reservation/ReservationController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.reservation.data.ReservationAddRequestDto;
-import roomescape.domain.reservation.data.ReservationResponseDto;
+import roomescape.dto.ReservationAddRequestDto;
+import roomescape.dto.ReservationResponseDto;
 import roomescape.domain.reservationtime.ReservationTimeService;
 import roomescape.domain.theme.ThemeService;
 import roomescape.entities.Reservation;
