+++ src/main/java/roomescape/domain/reservation/ReservationService.java
@@ -2,7 +2,7 @@ package roomescape.domain.reservation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import roomescape.domain.reservation.data.ReservationResponseDto;
+import roomescape.dto.ReservationResponseDto;
 import roomescape.entities.Reservation;
 import roomescape.errors.ErrorCode;
 import roomescape.exceptions.RoomEscapeException;
