+++ src/main/java/roomescape/domain/reservationtime/ReservationTimeController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.reservationtime.data.AvailableTimeResponse;
-import roomescape.domain.reservationtime.data.ReservationTimeAddRequestDto;
+import roomescape.dto.AvailableTimeResponse;
+import roomescape.dto.ReservationTimeAddRequestDto;
 import roomescape.entities.ReservationTime;
 
 import java.util.List;
