+++ src/main/java/roomescape/dto/AvailableTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservationtime.data;
+package roomescape.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
