+++ src/main/java/roomescape/dto/ReservationAddRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation.data;
+package roomescape.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
