+++ src/main/java/roomescape/dto/ReservationResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation.data;
+package roomescape.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
