+++ src/test/java/roomescape/MissionStepTest.java
@@ -5,9 +5,9 @@ import io.restassured.http.ContentType;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.domain.reservation.data.ReservationAddRequestDto;
-import roomescape.domain.reservationtime.data.ReservationTimeAddRequestDto;
-import roomescape.domain.theme.data.ThemeAddRequestDto;
+import roomescape.dto.ReservationAddRequestDto;
+import roomescape.dto.ReservationTimeAddRequestDto;
+import roomescape.dto.ThemeAddRequestDto;
 
 import java.time.LocalDate;
